requirements.txt
rank-bm25
sentence-transformers
numpy
scikit-learn
torch
faiss-cpu